Structure for events

+++ src/ORMMapViewer/Settings.cs
@@ -9,7 +9,7 @@ namespace ORMMap
         public readonly static LatLng startPosition = new LatLng(48.464717f, 35.046183f);
         public const uint renderDistanceX = 2;
         public const uint renderDistanceY = 2;
-        public const uint zoom = 13;
+        public const double zoom = 13;
         public const int cacheOpacity = 100;
     }
 }
